pyproject.toml
[tool.poetry]
name = "guiosoft-scripts"
version = "0.1.0"
description = "Automation scripts for every day"
authors = ["Guionardo <guionardo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "guiosoft_scripts"}]

[tool.poetry.dependencies]
python = "^3.10"
packaging = "^22.0"
requests = "^2.28.1"
tqdm = "^4.64.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
